Prevent yourself from deleting yourself
git-svn-id: file:///home/oestrich/tmp/img/trunk@537 ea04b438-0a61-0410-93d0-9bc465a3bab0

+++ admin/users/class.php
@@ -92,6 +92,9 @@ class Users extends lncln{
 	 * @param $id int User id to be deleted
 	 */
 	function deleteUser($id){
+		if($id = $this-user->userID)
+			return "";
+			
 		if(is_numeric($id)){
 			$sql = "DELETE FROM users WHERE id = " . $id;
 			mysql_query($sql);
